requirements.txt
tqdm==4.64.1
matplotlib==3.6.2  
chumpy==0.70 
opencv-python==4.6.0.66
pyyaml==6.0
transforms3d==0.4.1
trimesh==3.17.1
open3d==0.16.0
numpy==1.20.3
plyfile==0.7.4
scikit-image==0.19.3
tensorboardx==2.5.1